# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AutoFlow SaaS WhatsApp API
  version: 1.0.0
  description: API multiempresa para instancias, contatos, campanhas, mensagens e integracoes.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      security: []
      summary: Criar empresa e proprietario
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Register'}
      responses:
        '201': {$ref: '#/components/responses/Session'}
        '422': {$ref: '#/components/responses/Error'}
  /auth/login:
    post:
      security: []
      summary: Entrar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: {type: string, format: email}
                password: {type: string}
                tenantSlug: {type: string}
      responses:
        '200': {$ref: '#/components/responses/Session'}
        '401': {$ref: '#/components/responses/Error'}
...
  /instances:
    get:
      summary: Listar instancias da empresa
      responses:
        '200': {description: Lista de instancias}
    post:
      summary: Criar instancia respeitando o plano
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties: {name: {type: string, minLength: 2}}
      responses:
        '201': {description: Instancia criada}
        '409': {$ref: '#/components/responses/Error'}
...
  /contacts:
    get:
      summary: Listar contatos
      responses:
        '200': {description: Lista paginada}
    post:
      summary: Criar contato com origem de consentimento
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Contact'}
      responses:
        '201': {description: Contato criado}
...
  /messages:
    post:
      security: [{apiKeyAuth: []}]
      summary: Enviar mensagem direta idempotente
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: {type: string, minLength: 8}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instanceId, text]
              oneOf:
                - required: [contactId]
                - required: [phoneNumber]
              properties:
                instanceId: {type: string, format: uuid}
                contactId: {type: string, format: uuid}
                phoneNumber: {type: string, pattern: '^[1-9][0-9]{9,14}$', description: 'Celular internacional de um contato existente. Informe contactId ou phoneNumber.'}
                text: {type: string, minLength: 1, maxLength: 4096}
      responses:
        '202': {description: Mensagem enfileirada}
        '409': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: http
      scheme: bearer
      description: Chave wa_live com escopo apropriado.
...
  schemas:
    Register:
      type: object
      required: [companyName, ownerName, email, password, acceptedTermsVersion]
      properties:
        companyName: {type: string, minLength: 2}
        document: {type: string}
        timezone: {type: string, default: America/Sao_Paulo}
        ownerName: {type: string, minLength: 2}
        email: {type: string, format: email}
        password: {type: string, minLength: 10}
        acceptedTermsVersion: {type: string, example: '2026-07-16'}
    Contact:
      type: object
      required: [name, phoneNumber, consentStatus, consentSource]
      properties:
        name: {type: string}
        phoneNumber: {type: string}
        consentStatus: {enum: [UNKNOWN, GRANTED, REVOKED, BLOCKED]}
        consentSource: {type: string}
        consentAt: {type: string, format: date-time}
        consentEvidence: {type: string}
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: {type: string}
            message: {type: string}
        requestId: {type: string}
  responses:
    Session:
      description: Sessao criada
      content:
        application/json:
          schema:
            type: object
            properties:
              accessToken: {type: string}
              refreshToken: {type: string}
              expiresIn: {type: integer, example: 900}
    Error:
      description: Erro padronizado
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}